define intersection calculator as single module

+++ lab1/lines-app/src/main/java/mouse/univ/common/Numbers.java
@@ -2,7 +2,7 @@ package mouse.univ.common;
 
 public class Numbers {
     public static final double ACCURACY = 1e-8;
-    public static final double BOX_SIZE = 122;
+    public static final int BOX_SIZE = 122;
     public static boolean isZero(double number) {
         return Math.abs(number) < ACCURACY;
     }
